fix(blocks): use u64 UIDs — ROLI topology IDs exceed u32
ROLI topology UIDs like 17162900393818307115 overflow u32. Updated all
UID types to u64, binary frame size from 681 to 685 bytes (8-byte UID),
and pixel offset from 6 to 10.

+++ crates/hypercolor-core/src/device/blocks/backend.rs
@@ -25,7 +25,7 @@ pub struct BlocksBackend {
     /// Known devices reported by blocksd.
     devices: HashMap<DeviceId, BlocksDevice>,
     /// UID-to-`DeviceId` mapping for event routing.
-    uid_map: HashMap<u32, DeviceId>,
+    uid_map: HashMap<u64, DeviceId>,
     /// Per-device brightness (applied by blocksd).
     brightness: HashMap<DeviceId, u8>,
     /// Reconnection state.
@@ -33,7 +33,7 @@ pub struct BlocksBackend {
 }
 
 struct BlocksDevice {
-    uid: u32,
+    uid: u64,
     info: DeviceInfo,
     connected: bool,
     frames_sent: u64,
